+++ src/options.js
@@ -1,0 +1,26 @@
+/**
+ * Default options for lazy loading behavior.
+ *
+ * @type {Object}
+ *
+ * @param {string} [tag=selector] - The CSS selector for lazy loadable elements.
+ * @param {string} [toggleClass="-loaded"] - The class name to toggle on elements after loading.
+ * @param {Function} [onLoaded] - Callback function to execute when an element is successfully loaded.
+ * @param {Function} [onError] - Callback function to execute when an error occurs during loading.
+ * @param {Object} [observer={ root: null, threshold: 1, rootMargin: '300px 0px' }] - Configuration for IntersectionObserver used for lazy loading.
+ */
+const defaultOptions = {
+  tag: '[lazy]',
+  toggleClass: '-loaded',
+  onLoaded: () => {},
+  onError: (element, error) => {
+    console.error('🚀 Error on ~ element, error:', element, error);
+  },
+  observer: {
+    root: null,
+    threshold: 1,
+    rootMargin: '300px 0px',
+  },
+};
+
+export default defaultOptions;
