+++ test/index.html
@@ -39,17 +39,7 @@
 		<img lazy lazy-src="http://unsplash.it/400/400?random=7&gravity=center" alt="Random Image"/>
 		<img lazy lazy-src="http://unsplash.it/400/400?random=8&gravity=center" alt="Random Image"/>
 
-		<!-- library -->
-		<script src="/src/index.js"></script>
-
-		<!-- An Example -->
-		<script>
-			lazyLoad('[lazy]', {
-				onLoaded: (element) => {
-					console.log(element);
-				}
-			});
-		</script>
+		<script type="module" src="/test/index.js"></script>
 
 	</body>
 </html>
