+++ test/index.js
@@ -1,0 +1,8 @@
+import lazyLoad from '/src/index.js';
+
+lazyLoad({
+  tag: '[lazy]',
+  onLoaded: (element) => {
+    console.log(element);
+  },
+});
